Make history tests work, and minor tweaks to code
